move api calls in one file
